requirements.txt
streamlit
tesserocr
pillow
